Optimize sending data to server
Used cStringIO for the outgoing buffer and stored its data in reverse.  This
allows for truncating bytes sent off of the outgoing buffer rather than taking a
slice of the buffer which is very slow if there is a lot of data.  Other
manipulations of the buffer have been adapted to handle the data being in
reverse.